+++ src/FormilyDemo/components/CustomOnlyProp/index.ts
@@ -1,0 +1,5 @@
+import CustomOnlyProp from './CustomOnlyProp';
+
+export default CustomOnlyProp;
+
+export * from './CustomOnlyProp';