+++ src/FormilyDemo/components/ObservableDemo/index.ts
@@ -1,0 +1,5 @@
+import ObservableDemo from './ObservableDemo';
+
+export default ObservableDemo;
+
+export * from './ObservableDemo';