+++ src/PureHtml/EchartsMap/index.ts
@@ -1,0 +1,5 @@
+import EchartsMap from './EchartsMap';
+
+export default EchartsMap;
+
+export * from './EchartsMap';