+++ CHANGELOG.md
@@ -2,6 +2,39 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+## [1.3.0](https://github.com/tianxintiandisheng/demo-doc/compare/v1.1.9...v1.3.0) (2022-06-07)
+
+### ✏️ Documentation | 文档
+
+- **changelog:** changelog 格式化 ([1c8ac09](https://github.com/tianxintiandisheng/demo-doc/commit/1c8ac09ba75cd5c4cab39edfebe318afda26eeea))
+- **root:** 首页 md 修改 ([0d48b35](https://github.com/tianxintiandisheng/demo-doc/commit/0d48b350ff0c94a5bd25580c0f62f43a5ec0907a))
+- **root:** changelog 格式统一 ([f0c024d](https://github.com/tianxintiandisheng/demo-doc/commit/f0c024d7db29a3c64a59452345124d483c19d233))
+
+### 🐛 Bug Fixes | Bug 修复
+
+- **utils:** 浮点数的求和,sumFloatNum 方法不支持进位 ([baa55d9](https://github.com/tianxintiandisheng/demo-doc/commit/baa55d94fc898a649eb0f339cb0f5d567824c744))
+
+### 👷 Continuous Integration | CI 配置
+
+- **root:** 将更常用的文档打包命令设置为莫问的打包命令;组件打包命令修改; ([7991cc5](https://github.com/tianxintiandisheng/demo-doc/commit/7991cc524af94e1de41a08f9c7b46d5c407dbf92))
+- **root:** 添加自定义导航栏 ([63ca6cd](https://github.com/tianxintiandisheng/demo-doc/commit/63ca6cdf41307f5b8cb27f18cd1e90a24e3376e8))
+- **root:** 添加 html-loader 依赖,配置 html 文件的解析 ([652c855](https://github.com/tianxintiandisheng/demo-doc/commit/652c85587ad2a43f1116c79a0be60b4d0167b9b1))
+- **root:** changelog 隐藏 chore 信息,显示 ci 信息 ([11ceca6](https://github.com/tianxintiandisheng/demo-doc/commit/11ceca6bfb274e68062d9286d731a9d5c2a4f9e2))
+
+### ✨ Features | 新功能
+
+- **antd:** antd 弹出层的新建表单一个 bug 的 demo ([ad66248](https://github.com/tianxintiandisheng/demo-doc/commit/ad662481818a1b56c7fb86dad574da8123783ffc))
+- **formily:** 添加 Formily 表单设计器 基础 ([1e9e6e1](https://github.com/tianxintiandisheng/demo-doc/commit/1e9e6e1e48b8de79e6f122c678b823730769670e))
+- **formily:** 添加 Formily 的基础组件 ([641737d](https://github.com/tianxintiandisheng/demo-doc/commit/641737dc9c36a9356dd4c44eec7eb891d0ee53fa))
+- **formily:** 自定义组件-属性映射 ([0d159dd](https://github.com/tianxintiandisheng/demo-doc/commit/0d159dd9d81f59490efad34b4b70ee457fa5801a))
+- **formily:** formily observable demo ([d1eeeae](https://github.com/tianxintiandisheng/demo-doc/commit/d1eeeae874d9ef4a414fdb08d9ff49b432a453ce))
+- **root:** 添加 Formily 表单设计器 相关依赖 ([18db006](https://github.com/tianxintiandisheng/demo-doc/commit/18db006b877dda174078804905c10d4354a5643d))
+- **root:** 添加 antd；antd 样式按需加载 ([b2feef4](https://github.com/tianxintiandisheng/demo-doc/commit/b2feef41d1118362754e10e6ae573f7717b371f5))
+- **root:** 引入 formly 的相关依赖 ([92862fd](https://github.com/tianxintiandisheng/demo-doc/commit/92862fdf3d31b38235e07bfe7421335bc57cc968))
+- **tool:** 基于 Echarts 的中国地图 ([656e81a](https://github.com/tianxintiandisheng/demo-doc/commit/656e81ad6ad292d95abe5267a44ccff74c0919a3))
+- **utils:** floatNumSum 添加 number-precision 的计算结果 ([cc09575](https://github.com/tianxintiandisheng/demo-doc/commit/cc09575896f6bce5ed286c59ad26ef5a63c776dd))
+- **utils:** js 中浮点数的计算（求和） ([e22b8b5](https://github.com/tianxintiandisheng/demo-doc/commit/e22b8b5589aa1cf57952a746c62dc05f77d83be7))
+
 ## [1.2.0](https://github.com/tianxintiandisheng/demo-doc/compare/v1.1.9...v1.2.0) (2022-05-09)
 
 ### 👷 Continuous Integration | CI 配置